tidy projects logic
