+++ src/App.tsx
@@ -1,15 +1,15 @@
-import './App.css'
-import '@radix-ui/themes/styles.css'
-import { Theme } from '@radix-ui/themes'
-import { BrowserRouter as Router, Routes, Route } from 'react-router-dom'
-import NavBar from './components/NavBar'
-import Work from './components/Work'
-import About from './components/About'
-import HomeHero from './components/HomeHero'
-import ProjectsPage from './components/ProjectsPage'
+import './App.css';
+import '@radix-ui/themes/styles.css';
+import { Theme } from '@radix-ui/themes';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import NavBar from './components/NavBar';
+import Work from './components/Work';
+import About from './components/About';
+import HomeHero from './components/HomeHero';
+import ProjectsPage from './components/Projects/ProjectsPage';
 
 function Home() {
-  return <HomeHero />
+  return <HomeHero />;
 }
 
 function App() {
@@ -25,7 +25,7 @@ function App() {
         </Routes>
       </Router>
     </Theme>
-  )
+  );
 }
 
-export default App
+export default App;
