+++ src/components/Projects/projects.ts
@@ -1,0 +1,73 @@
+import { ProjectsType } from './../../types/types';
+export const projects: ProjectsType[] = [
+  {
+    id: 'assistant',
+    title: 'Assistant',
+    period: 'March 2024',
+    description: [
+      'Developed a Python-based wellness assistant that integrates with Oura Ring to track sleep patterns and provide personalized health insights. ',
+      'Implemented automated scheduling system for daily workouts, affirmations, and motivational content delivery. ',
+      'Built SMS notification system using TextBelt API for timely updates and reminders. ',
+    ],
+    technologies: [
+      'Python',
+      'Oura Ring API',
+      'TextBelt API',
+      'Environment Variables',
+      'Scheduling System',
+    ],
+    githubUrl: 'https://github.com/ocjarman/assistant.git',
+    imageUrl: '/portfolio/assistant.jpg',
+  },
+  {
+    id: 'codenames',
+    title: 'Codenames',
+    period: 'February 2023',
+    description: [
+      'Built and deployed a multi-player Codenames board game with a team of developers at Fullstack Academy of Code. ',
+      'Utilized a combination of Firebase and Redux to listen to and update game state dependent on players turn and changing game statuses. ',
+      'Constructed a backend API to hold gameboard data and to secure game answers away from players. ',
+    ],
+    technologies: [
+      'TypeScript',
+      'React',
+      'Redux',
+      'Firebase',
+      'Node',
+      'Express',
+      'PostgreSQL',
+      'Sequelize',
+      'CSS',
+      'Render',
+    ],
+    githubUrl: 'https://github.com/2208-Capstone-Team-2/codenames',
+    liveUrl: 'https://codenames-0nt7.onrender.com/',
+    imageUrl: '/portfolio/codenames.jpg',
+  },
+  {
+    id: 'rrr',
+    title: 'Right Round Records',
+    period: 'November 2022',
+    description: [
+      'Built and deployed an e-commerce site that lets users view and checkout record albums.',
+      'Structured various database models for inventory, orders, and applied secure routes for adding, editing, and deleting shopping cart items. ',
+      'Utilized OAuth for user authentication and Material-UI for styling. ',
+    ],
+    technologies: [
+      'OAuth',
+      'React',
+      'Redux',
+      'Node',
+      'Express',
+      'PostgreSQL',
+      'Sequelize',
+      'CSS',
+      'HTML',
+      'Material-UI',
+      'Discogs API',
+    ],
+    githubUrl: 'https://github.com/2208-GS-Team-3/grace-shopper-team-three',
+    liveUrl: 'https://rightroundrecords.us',
+    imageUrl: '/portfolio/rightround.jpg',
+  },
+];
