+++ src/types/types.ts
@@ -1,0 +1,11 @@
+export type ProjectsType = {
+  id: string;
+  title: string;
+  period: string;
+  description: string[];
+  technologies: string[];
+  githubUrl: string;
+  liveUrl?: string;
+  imageUrl?: string;
+  backgroundColor?: string;
+};
